requirements.txt
absl-py==2.1.0
accelerate==0.27.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
bigframes==2.8.0
bitsandbytes==0.43.2
certifi==2025.6.15
cesium==0.12.4
charset-normalizer==3.4.2
chex==0.1.9
chroma-hnswlib==0.8.3
chromadb==0.5.5
click==8.1.7
cloud-tpu-client==0.10
cloudpickle==3.0.0
colorama==0.4.6
colossalai==0.4.1
conda-package-handling==2.4.0
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.6.0
dataproc-spark-connect==0.7.5
db-dtypes==1.2.0
distrax==0.1.5
dm-tree==0.1.8
etils==1.8.0
fastapi==0.112.0
filelock==3.18.0
flatbuffers==25.2.22
flax==0.9.1
fonttools==4.56.0
frozenlist==1.7.0
fsspec==2025.5.1
gast==0.6.0
gcsfs==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==1.34.1
google-api-python-client==2.146.0
google-auth==2.40.3
google-auth-httplib2==0.2.2
google-auth-oauthlib==1.2.3
google-cloud-bigquery==3.25.0
google-cloud-bigquery-connection==1.12.2
google-cloud-core==2.4.1
google-cloud-datastore==2.18.0
google-cloud-functions==1.14.0
google-cloud-iam==2.17.0
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.19.0
google-colab==1.0.0
google-crc32c==1.5.0
google-generativeai==0.8.1
google-pasta==0.2.0
googleapis-common-protos==1.63.2
greenlet==3.2.3
grpcio==1.66.1
grpcio-status==1.65.1
h11==0.16.0
h5py==3.11.0
hf-xet==1.1.5
holidays==0.59
httpcore==1.0.9
httplib2==0.24.2
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.24.2
idna==3.10
intel-cmplr-lib-rt==2024.2.0
intel-cmplr-lib-ur==2024.2.0
intel-openmp==2024.2.0
jax==0.4.32
jax-cuda12-pjrt==0.4.32
jax-cuda12-plugin==0.4.32
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2.3.1
jupyter-client==8.6.3
jupyter-console==6.6.4
jupyter-core==5.7.2
jupyter-events==0.10.1
jupyter-kernel-gateway==2.5.2
jupyter-lsp==2.3.0
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyter-ydoc==0.3.1
jupyterlab==4.3.1
jupyterlab-git==0.52.0
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.12
keras==3.6.0
kiwisolver==1.4.5
langchain==0.3.27
langchain-chroma==0.1.2
langchain-community==0.3.28
langchain-core==0.3.75
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.9
langchainhub==0.1.21
langsmith==0.4.1
libclang==18.1.1
llvmlite==0.43.0
lvis==0.5.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.7.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==0.8.4
mizani==0.11.7
mkl==2025.2.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.1
mkl_umath==0.1.1
ml-dtypes==0.3.2
mlxtend==0.23.4
mpmath==1.3.0
multidict==6.6.3
mypy-extensions==1.1.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==6.5.4
notebook-shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvbackend==0.6.0
nvidia-nvinfer==8.6.1
nvidia-nvinfer-plugin==8.6.1
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.16.1
onnxruntime==1.18.1
opencv-python==4.10.0.84
opt-einsum==3.3.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
optax==0.2.4
orbax-checkpoint==0.5.21
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandas-gbq==0.29.1
pandas-stubs==2.2.3.240715
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==11.2.1
pip==24.2
platformdirs==4.2.2
plotnine==0.14.5
plotly==5.25.1
posthog==3.5.0
prometheus-client==0.20.0
promise==2.3
prompt-toolkit==3.0.47
propcache==0.3.2
proto-plus==1.24.0
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pulsar-client==3.6.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydeck==0.9.1
Pygments==2.19.0
PyJWT==2.8.0
pypika==0.48.9
pyparsing==3.1.2
python-dateutil==2.9.0
python-dotenv==1.0.1
python-json-logger==2.1.0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.36.1
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.1
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.19.1
rsa==4.9
safetensors==0.4.4
scikit-image==0.25.0
scikit-learn==1.2.2
scipy==1.15.3
Send2Trash==1.8.3
sentence-transformers==2.7.0
sentencepiece==0.2.0
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.37.2
statsmodels==0.14.3
sympy==1.13.1
tbb==2022.2.0
tbb4py==2022.2.0
tcmlib==1.4.0
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-datasets==4.9.6
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-metadata==1.15.0
tensorflow-probability==0.24.0
tensorflow-text==2.18.0
termcolor==2.4.0
terminado==0.18.1
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2024.8.12
tinycss2==1.3.0
tokenizers==0.19.1
toolz==0.12.1
torch==2.5.0
torchvision==0.18.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.38.2
triton==2.5.0
types-pytz==2024.1.0.20240417
types-requests==2.32.4.20250809
typing-extensions==4.14.0
typing-inspect==0.9.0
tzdata==2024.1
ultralytics==8.2.71
urllib3==2.5.0
uvicorn==0.30.5
wcwidth==0.2.13
webcolors==2.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.12
wrapt==1.16.0
yarl==1.20.1
yfinance==0.2.46
ypy-websocket==0.10.1
zstandard==0.23.0